moved WebDataset test file creation to test file
